ELEC-43 Resolved tituschow's comments
