+++ libraries/ms-common/src/stm32f0xx/critical_section.c
@@ -1,0 +1,22 @@
+#include "critical_section.h"
+
+#include <stdbool.h>
+
+#include "interrupt.h"
+#include "stm32f0xx_misc.h"
+
+bool critical_section_start(void) {
+  if (!__get_PRIMASK()) {
+    __disable_irq();
+    // Interrupts got disabled.
+    return true;
+  }
+  // Interrupts did not get disabled.
+  return false;
+}
+
+void critical_section_end(bool disabled_in_scope) {
+  if (__get_PRIMASK() && disabled_in_scope) {
+    __enable_irq();
+  }
+}
