+++ platform/stm32f0xx/platform.mk
@@ -22,7 +22,7 @@ LDSCRIPT_DIR := $(PLATFORM_DIR)/ldscripts
 SCRIPT_DIR := $(PLATFORM_DIR)/scripts
 
 # Build flags for the device
-CDEFINES := USE_STDPERIPH_DRIVER STM32F072 SEMIHOSTING_ENABLED
+CDEFINES := USE_STDPERIPH_DRIVER STM32F072 SEMIHOSTING_ENABLED=1
 CFLAGS := -Wall -Werror -g -Os -std=c99 -Wno-unused-variable -pedantic \
           -ffunction-sections -fdata-sections -fno-builtin -flto \
           --specs=nosys.specs --specs=nano.specs \
